[marge]マージ作業とタツキのスクリプト修正。
