+++ .gitignore
@@ -373,3 +373,4 @@ Assets/AurynSky/Dungeon Pack/Textures/CubeParticleDungeon.psd.meta
 Assets/AurynSky/Dungeon Pack/Textures/Water.png
 Assets/AurynSky/Dungeon Pack/Textures/Water.png.meta
 Assets/TextMesh Pro/
+Assets/Ignore/
