+++ Assets/Ignore.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 057d5de263ca4184cb0d189627004151
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
