+++ Assets/Ingame/Title.unity
@@ -3964,7 +3964,7 @@ MonoBehaviour:
   m_text: MOVERS.INC
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: 20700f5492ca3a544bf924c3ab381f79, type: 2}
-  m_sharedMaterial: {fileID: 2100000, guid: 0a23c38865136fa44beaffa5bb65125c, type: 2}
+  m_sharedMaterial: {fileID: 3974983726989887922, guid: 20700f5492ca3a544bf924c3ab381f79, type: 2}
   m_fontSharedMaterials: []
   m_fontMaterial: {fileID: 0}
   m_fontMaterials: []
@@ -4484,6 +4484,10 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 1237958953}
     m_Modifications:
+    - target: {fileID: 319844722129665703, guid: e0c557f834b53d34da8869729c004652, type: 3}
+      propertyPath: nextSceneName
+      value: Lobby
+      objectReference: {fileID: 0}
     - target: {fileID: 2715918592479864133, guid: e0c557f834b53d34da8869729c004652, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
