+++ Assets/_Higashi/CommonSliders.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 39306e434cd370f41b1a0a1e4ca57a0c
