+++ Assets/_Higashi/UIPrefab/UI Manager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff9f4e10d4ef3d04a801e1e4f96a628e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
