+++ Assets/_Tatsuki/CollectionArea.cs
@@ -27,7 +27,7 @@ public class CollectionArea : MonoBehaviour
         if (other.CompareTag("Luggage"))
         {
             var luggage = other.gameObject.GetComponent<Luggage>();
-            _scoreManager.SetScore(luggage.Score);
+           // _scoreManager.SetScore(luggage.Score);
           //  _scoreManager.SetText(_scoreManager.NowScore.ToString());
             OnEnter?.Invoke(other.gameObject);
             if (luggage.Target == Target.Target) OnEnterLuggage?.Invoke(1);
@@ -42,7 +42,7 @@ public class CollectionArea : MonoBehaviour
         if (other.CompareTag("Luggage"))
         {
             var luggage = other.gameObject.GetComponent<Luggage>();
-            _scoreManager.SetScore(-luggage.MaxScore);
+            //_scoreManager.SetScore(-luggage.MaxScore);
             luggage.MaxScore = luggage.Score;
           //  _scoreManager.SetText(_scoreManager.NowScore.ToString());
             OnExit?.Invoke(other.gameObject);
