+++ Assets/_Tatsuki/LuggageCollector.cs
@@ -20,6 +20,7 @@ public class LuggageCollector : MonoBehaviour
         {
             collectedCount++;
             toRemove.Add(item);
+            _scoreManager.SetScore(item.GetComponent<Luggage>().Score);
             Destroy(item);
         }
